New m_hex, m_count, m_begin and m_end macros
`rbp` now marks the base of the stack and so can be used
to calculate the size of the stack. There is now an `m_count`
macro which will push the number of elements on the stack.

Introduced `m_exit` macro to call the exit syscall with
a particular status code.

`m_hex` macro to print a quad word to stdout in hexadecimal
format. It currently prints the value in reverse but I'm
looking into concise ways to print the bytes in the correct
order.

Add `m_begin` and `m_end` macros to take care of boilerplate
for setting up and tearing down the program.