Preprocessor script and fixes
There is now a preprocessor script that converts what looks
like more traditional concatenative code into macro calls
which should make the development process slightly easier.
This script is unbelievably error prone and good luck fixing
any errors you do encounter because its all just textual
substitution. The upside to all of this is that when time
comes to bootstrap the compiler, it being written in a stack
based style originally should make it fairly straightforward
to port over.

Some fixes have been implemented for m_swap, m_abs and m_hex.
m_swap was looking in the wrong direction for the element so
the sign has been reversed. m_abs was using some old copy/pasted
code that was irrelevant to its actual function so it's now been
rewritten to do what it's supposed to do. m_hex had an issue with
mapping numeric digits and so now it prints nibbles correctly.