Debugging symbols, operand reversal and some fixes
Add a flag for emitting debug symbols from NASM and preventing
stripping by ld. This makes things a lot easier to debug with
gdb.

Removed m_fn_begin and m_fn_end. They were fairly redundant
considering m_fn_begin is currently empty and m_fn_end is just
a synonym for m_go. Maybe they'll have use in the future but for
now they actually get in the way moreso than anything.

All binary operators now reverse the operands before peforming the
operation. As an example, take the code `10 5 -`. In a language like
FORTH, this would result in `10 - 5 = 5` but in deck this would have
resulted in `5 - 10 = -5`. I find that the FORTH order of operands
is more readable but it does incur some extra overhead from the swapping
of operand order. This might prove to be a bad decision that results in
a roll back but only time will tell.

Fixed an issue with the logical operators using `test` rather than `cmp`.
`test` is essentially a bitwise AND so it doesn't actually work for anything
other than AND in deck.

The operand order has been changed around for all operators that accept
a condition value. This makes conditions a lot more readable and results
in less stack juggling. Another bonus is that the implementation for
m_choose has been vastly simplified.

m_here has been fixed because previously it was untested and would produce
an error when compiled under NASM. NASM apparently doesn't allow you to
push the instruction pointer to the stack directly, you have to use a
combination of `lea` with `rel` and an offset of `0`. Now, I'm not sure if
this is NASM just working with a quirk of x86 but it's pretty weird. In any
case, m_here now works correctly.

m_hex now prints a newline.

m_header no longer jumps to `main`, instead it just continues on through
to whatever code exists at the top of the source file.