Split up macros into multiple files
The previous macros.asm file has been split up and compartmentalised
into a number of files for better organisation. Some macros have been
turned into functions.

The naming scheme for macros has been changed from a leading `m_` to
triple leading `___`. All builtins are now prefixed with this. Functions
are instead prefixed with their semantic group (i.e. `io`, `sys` etc.).

A number of macros have been added: ___call, ___rotl, ___rotr, ___at.

It is now possible to call a function using the syntax `&<name>` as
opposed to first pushing the return address and using `.` manually.
This should simplify cases where you simply want to return to the next
instruction after a call but the underlying mechanism is still the exact
same as before.

A basic highlighter for Kakoune has been added. It has a few issues but
will improve the more I use it.

The previously named `hex` macro (now a function called `io_hexln`) was
clearly broken when viewed with a hex viewer (like `xxd`). Null bytes
were being written after every string incorrectly. The source of this bug
is x86-64 not allowing for the pushing of single bytes to the stack so
every value we pushed was implicitly being upcasted to a 16 bit word. This
has now been fixed by manually decrementing `rsp` and storing the byte to
the top of the stack.