Deque functionality and some new builtins/functions
Add deque functionality which allows pushing popping to the
opposite end of the stack. This effectively gives you two stacks
to work with and vastly simplifies some algorithms. I believe this
now makes Deck undeniably Turing complete because you can model a
tape as two stacks to the left and right of the read/write head of
a Turing machine. It's possible it was already Turing complete through
use of the get/set builtins (formerly `at`).

The deque currently reserves 100 qwords but ideally I would like to split
the stack in half by finding out the stack size either at runtime or
compile time.

There is still one issue present with the deque functionality when you
underflow the stack and start eating into elements pushed to the other
side of the stack. `rax` isn't updated to take into account when `rbp`
is greater than `rsp` but maybe there's an easy solution for this I can
try to find.

Added two new builtins `mget` and `mset` for working with arbitrary
pointers to memory. You can use this when working with the stack but
it's not recommended due to the accumulator behaviour of the top of
stack value. This should prove useful when dynamic memory allocation
is eventually implemented.

The Deck highlighter for Kakoune has been updated to reflect some
changes to builtin names and new additions.

A new IO routine has been added: `io_intln` which prints the top of
the stack to stdout in base-10/decimal format.