Fix zero value being pushed at first call to ___push
Previously the first call to ___push in the program would move
the value `0` in register `rax` to the actual stack causing an
issue where utilising the stack as a deque would have a `0`
separating the front and back halves of the deque. This was
mentioned in a previous commit.

The fix is to find and replace the first call to `___push` in the
program with a call to a new macro `___first_push` which avoids
saving the existing value in `rax`.