Stack markers and simplification of assembly routines/macros + introduction of many new builtins
Assembly routines and macros no longer use `rax` as the top-of-stack.
This was causing a number of issues when using builtins like count (`#`)
or when the stack was empty and you wanted to underflow the stack to reach
elements on the other side of the deque. My logic in removing what was a
fairly big optimisation is that it's better to suffer the hit now in favour
of simplicity and getting things working and then go back and improve the
situation later when we have a self-hosting and optimising compiler.

The operand order for binary operators has been reverted because I think it
fits more naturally with the RPN style. Yes I realise I'm going against
pretty much every precedent here but so what?

Renamed some of the core macros to remove the `___res_` prefix which was a
bit unnecessary given we have the `___` prefix already.

Removed the program footer because it wasn't really serving much purpose
other than to bloat the executable. If the user wishes to exit gracefully,
let them do it themselves rather than forcing the decision upon them.

Functions no longer implicitly push/pop from the deque with return addresses
nor do they set the marker implicitly. Everything is now in full view of the
user. Yes, this does mean slightly more verbosity but it also means more
control and that is essentially the whole point of this project.

`___mark` and `___unmark` have been rewritten to save/restore the base
pointer from the deque (which we use somewhat as a return stack). We now
also give them the sigils `[` and `]`. This might be one of the biggest and
best introductions to the language so far because it means you can isolate
values from the stack for variadic functions like `sum` or `dbg` which consume
the entire stack.

Introduce `___call_go` builtin to complement the `___call` builtin. `___call`
works by taking a _compile-time_ value to jump to whereas `___call_go` takes
the address from the top of the stack at runtime.

Introduce `___branch` which works the same way as `___choose` but jumps to
the evaluated value instead of just reducing to it.

`___tuck` builtin to push a copy of the top-of-stack underneat the second
element.

Calling functions is now just a matter of naming the function instead of before
where you had to prefix the name with `&`. This makes the dichotomy of builtins
and functions less noticeable since they are called the same way now.

Updated the the Kakoune highlighter significantly. Everything is now highlighted
properly and it looks pretty good in my editor.

Updated the future grammar of Deck a little bit.

`sys_ok` function to exit with `0` status.