Minor changes + Experimentation
Add a short factorial implementation to the README to show off
a bit of Deck.

Renamed `&.` to `call`. This builtin is used to call a function
whose address is on top of the stack while also pushing a
continuation address to the stack.

Removed the `drop` alias. Just use `pop`.

Added a bunch of functions to the `deck.dk` file. This is mostly
serving as a playground for implementing things in Deck at the
moment and is pretty rough and disorganised. Eventually I'll get
around to adding these functions to their own file.

Add some new operators/keywords to the Kakoune highlighter. Namely
the `[` and `]` marker operators and `br`, `call`, `mget` & `mset`.

Renamed some things in the grammar file aswell as a new syntax for
function definitions.