Ported some builtins to Deck and implemented basic memory allocation functions
Some of the builtins have been ported into Deck itself: `swp`, `rotr`
and `clear`.

Add two new builtins: `word` and `byte` which give some information
on the size of values on the current platform. Using these gives code
some amount of portability across architectures with different word
sizes.

Some new functions for working with pointers: `ptr++`, `ptr--` and `cell`.

Switched the other of modifier characters for function names. Previously
we used prefix modifiers but now they're postfix. I just think it fits
better. This is all purely convention but it's good to be consistent.

Add some documentation to some of the more complicated functions in the
prelude.

`times` function to repeat something N times. Very handy for writing for
loops without all of the surrounding boilerplate.

Add the beginnings of memory allocation with `sys_mmap` and `sys_unmap`.
These functions are supposed to serve as very low level facilities from
which to build more convenient functions on top and in Deck.

Add the `m4` preprocessor to the compiler pipeline. This allows you to
define things and include other files.