Fix sys_mmap function
sys_mmap was not saving the `r9` register before performing the
syscall and so it would be set to some garbage value and cause some
hard to diagnose issues somewhere else in the code. The fix is to
save `r9` to the stack and restore it at the end of the function before
returning.