Function reordering
Add a new pass to reorder functions in order of dependencies.
This is done by doing a DFS on functions and references until
we reach the leaves and then as we unwind the stack, we build
up a new tree with the functions in the right order. You can
think of this as a topological sort on a call-graph.

This should make future passes a bit easier to write since we
wont need to do any extra bookkeeping for out of order
references.

In addition to the above feature, I've split up the existing
passes into their own seperate files for organisation sake
because it was becoming a bit unwieldy to have them all stuffed
in the same source file.

Some "utility" passes were introduced to help with the
implementation of function reordering: `ranges` and `refgraph`.

`ranges` records the beginning and ending iterators for every
function to make it easy to copy function bodies around.

`refgraph` generates a reference graph which _was_ previously
used when prototyping the reordering pass but is actually no
longer used. I might end up removing it but it could turn out
to be useful information to have for future passes.